fastapi/requirements.txt
fastapi
uvicorn[standard]
httpx
asyncpg
python-multipart
pydantic
pydantic-settings
bcrypt
python-jose[cryptography]
passlib[bcrypt]
websockets
python-dotenv
reportlab
Pillow
